consistency

+++ Codeforces/evenodds.java
@@ -1,18 +1,17 @@
-import java.util.Scanner;
+import java.util.*;
 
 public class evenodds {
-     public static void main(String[] args) {
+    public static void main(String[] args) {
         Scanner sc = new Scanner(System.in);
 
-        int n = sc.nextInt();
-        int k = sc.nextInt();
+        long n = sc.nextLong();
+        long k = sc.nextLong();
 
-        int oddCount = (n + 1) / 2;
+        long oddCount = (n + 1) / 2;
 
-        if (k <= oddCount) {
+        if(k <= oddCount){
             System.out.println(2 * k - 1);
-        } 
-        else {
+        } else {
             System.out.println(2 * (k - oddCount));
         }
     }
